# src/main/java/com/example/demo/model/dto/PhoneDto.java
package com.example.demo.model.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Setter
@Getter
@AllArgsConstructor
@NoArgsConstructor
public class PhoneDto {

    private String number;
    private String cityCode;
    private String countryCode;

}